🐛 fix: error en el build por los componentes de landing en la propia page
